kvmctl protocol markdown doc
Protocol description for the tesmart kvm is here, better formatted and doesn't consume as much space in the code as the previous comments did. As this documentation did nothing to actually describe what the code does other than explain the significance of the bytes sent over the wire to the KVM a separate document seemed appropriate.